right_hand_plan/requirements.txt
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
boto3==1.28.62
PyPDF2==3.0.1
reportlab==4.0.4
python-multipart==0.0.6